Time: 9 ms (14.40%) | Memory: 47.2 MB (23.84%) - LeetSync

+++ 981-delete-columns-to-make-sorted/delete-columns-to-make-sorted.java
@@ -4,7 +4,7 @@ class Solution {
         int count = 0;
         for(int i=0;i<strs[0].length();i++){
             for(int j=0;j<strs.length-1;j++){
-                if((int)strs[j].charAt(i) <= (int)strs[j+1].charAt(i)){
+                if(strs[j].charAt(i) <= strs[j+1].charAt(i)){
                     count++;
                 }
             }
